+++ Program/Mesh/material.cpp
@@ -19,7 +19,7 @@ QDataStream& operator >> (QDataStream& out, Material& m) {
     out >> type;
     m.type = static_cast<Material::Type>(type);
 
-    size_t size;
+    quint32 size;
     out >> size;
     m.values.resize(size);
     for (size_t i = 0; i < size; ++i) {
@@ -30,7 +30,7 @@ QDataStream& operator >> (QDataStream& out, Material& m) {
 }
 QDataStream& operator << (QDataStream& in, const Material& m) {
     in << static_cast<int>(m.type);
-    size_t size = m.values.size();
+    quint32 size = m.values.size();
     in << size;
     for (size_t i = 0; i < size; ++i ) {
         in << m.values[i];
